requirements.txt
flask==2.0.1
sqlalchemy==1.4.23
elasticsearch==7.15.0
pytest==6.2.5
pillow==8.3.2
python-dotenv==0.19.0
gunicorn==20.1.0
requests==2.26.0 